added test for db. Moved config.go to utils for tests [MCORE-5]
