+++ server.go
@@ -2,13 +2,14 @@ package main
 
 import (
 	"github.com/Schtolc/mooncore/handlers"
+	"github.com/Schtolc/mooncore/utils"
 	"github.com/jinzhu/gorm"
 	"github.com/labstack/echo"
 	"github.com/labstack/echo/middleware"
 )
 
 // InitServer inits echo server: sets access logs and handlers
-func InitServer(config Config, db *gorm.DB) (e *echo.Echo) {
+func InitServer(config utils.Config, db *gorm.DB) (e *echo.Echo) {
 	server := echo.New()
 	server.Use(middleware.LoggerWithConfig(GetAccessConfig(config.Logs.Access)))
