+++ utils/config.go
@@ -1,9 +1,19 @@
-package main
+package utils
 
 import (
 	"github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v2"
 	"io/ioutil"
+	"path/filepath"
+	"runtime"
+)
+
+var (
+	// This is a way to get path of the current file in runtime. It's nessary,
+	// because binary and tests run from different dirs and we need to find config file
+	// by relative path in both cases. Taken from https://stackoverflow.com/a/38644571
+	_, b, _, _  = runtime.Caller(0)
+	projectRoot = filepath.Join(filepath.Dir(b), "..")
 )
 
 // Config is in-code representation of config.yml.
@@ -28,7 +38,7 @@ type Config struct {
 
 // GetConfig reads config.yml and return filled Config struct. If any error occurs program is terminated.
 func GetConfig() Config {
-	content, err := ioutil.ReadFile("config.yml")
+	content, err := ioutil.ReadFile(filepath.Join(projectRoot, "config.yml"))
 	if err != nil {
 		logrus.Fatal(err)
 	}
